+++ src/main/java/cart/WebMvcConfig.java
@@ -1,7 +1,7 @@
 package cart;
 
 import cart.dao.MemberDao;
-import cart.ui.MemberArgumentResolver;
+import cart.controller.MemberArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
