+++ src/main/java/cart/controller/ProductApiController.java
@@ -1,6 +1,6 @@
-package cart.ui;
+package cart.controller;
 
-import cart.application.ProductService;
+import cart.service.ProductService;
 import cart.dto.ProductRequest;
 import cart.dto.ProductResponse;
 import org.springframework.http.HttpStatus;
